+++ FrameWork Ver3.1/Base/Source/Tower.h
@@ -36,10 +36,10 @@ public:
 
 	STRATEGY strategy;
 	void SetType(GEOMETRY_TYPE meshID);
-	void SetCost(float c);
-	float GetCost();
-	void SetAtkDmg(float ad);
-	float GetAtkDmg();
+	void SetCost(int c);
+	int GetCost();
+	void SetAtkDmg(int ad);
+	int GetAtkDmg();
 	void SetSpdRate(float sr);
 	float GetSpdRate();
 	void SetRange(float r);
@@ -70,13 +70,13 @@ public:
 
 	string upgrades[2];
 
-	float atkDamage;
+	int atkDamage;
 	float atkSpeed;
 	float atkRange;
 	int buffCounter;
 
 	ESSENCE_TYPE essence;
-	float essenceUpgradeCost;
+	int essenceUpgradeCost;
 private:
 
 protected:
@@ -87,7 +87,7 @@ protected:
 	
 	float p_frequency;
 	float p_spawnTimer;
-	float towerUpgradeCost;
+	int towerUpgradeCost;
 	
 
 	float p_speed;
